+++ R/tm_g_pp_therapy.R
@@ -114,7 +114,9 @@ template_therapy <- function(dataname = "ANL",
   plot_list <- add_expr(
     list(),
     substitute(expr = {
-      max_day <- max(dataname[[cmendy_char]], na.rm = TRUE)
+      dataname[[cmstdy_char]] <- as.numeric(dataname[[cmstdy_char]])
+      dataname[[cmendy_char]] <- as.numeric(dataname[[cmendy_char]])
+      max_day <- max(dataname[[cmendy_char]], na.rm = T)
       data <- dataname %>%
         dplyr::filter(atirel %in% c("CONCOMITANT", "PRIOR")) %>% # remove PRIOR_CONCOMITANT
         dplyr::select_at(cols_to_include) %>%
@@ -238,6 +240,7 @@ template_therapy <- function(dataname = "ANL",
 #'
 #' @examples
 #' library(scda)
+#'
 #' ADSL <- synthetic_cdisc_data("latest")$adsl
 #' ADCM <- synthetic_cdisc_data("latest")$adcm
 #'
@@ -248,16 +251,10 @@ template_therapy <- function(dataname = "ANL",
 #' ADCM$CMDOSU <- "U"
 #' ADCM$CMROUTE <- "CMROUTE"
 #' ADCM$CMDOSFRQ <- "CMDOSFRQ"
-#' ADCM$CMSTDY <- 1
-#' ADCM[ADCM$CMCAT == "medcl B", ]$CMSTDY <- 20
-#' ADCM[ADCM$CMCAT == "medcl C", ]$CMSTDY <- 150
-#' ADCM$CMENDY <- 500
-#' ADCM[ADCM$CMCAT == "medcl B", ]$CMENDY <- 700
-#' ADCM[ADCM$CMCAT == "medcl C", ]$CMENDY <- 1000
 #' ADCM$CMASTDTM <- ADCM$ASTDTM
 #' ADCM$CMAENDTM <- ADCM$AENDTM
 #' rtables::var_labels(
-#'   ADCM[c("CMINDC", "CMTRT", "CMSTDY", "CMENDY")]
+#'   ADCM[c("CMINDC", "CMTRT", "ASTDY", "AENDY")]
 #' ) <- c(
 #'   "Indication",
 #'   "Reported Name of Drug, Med, or Therapy",
@@ -277,16 +274,10 @@ template_therapy <- function(dataname = "ANL",
 #'       ADCM$CMDOSU <- "U"
 #'       ADCM$CMROUTE <- "CMROUTE"
 #'       ADCM$CMDOSFRQ <- "CMDOSFRQ"
-#'       ADCM$CMSTDY <- 1
-#'       ADCM[ADCM$CMCAT == "medcl B", ]$CMSTDY <- 20
-#'       ADCM[ADCM$CMCAT == "medcl C", ]$CMSTDY <- 150
-#'       ADCM$CMENDY <- 500
-#'       ADCM[ADCM$CMCAT == "medcl B", ]$CMENDY <- 700
-#'       ADCM[ADCM$CMCAT == "medcl C", ]$CMENDY <- 1000
 #'       ADCM$CMASTDTM <- ADCM$ASTDTM
 #'       ADCM$CMAENDTM <- ADCM$AENDTM
 #'       rtables::var_labels(
-#'         ADCM[c("CMINDC", "CMTRT", "CMSTDY", "CMENDY")]) <- c(
+#'         ADCM[c("CMINDC", "CMTRT", "ASTDY", "AENDY")]) <- c(
 #'           "Indication",
 #'           "Reported Name of Drug, Med, or Therapy",
 #'           "Study Day of Start of Medication",
@@ -335,16 +326,17 @@ template_therapy <- function(dataname = "ANL",
 #'         selected = "CMDOSFRQ"
 #'       ),
 #'       cmstdy = choices_selected(
-#'         choices = variable_choices(ADCM, "CMSTDY"),
-#'         selected = "CMSTDY"
+#'         choices = variable_choices(ADCM, "ASTDY"),
+#'         selected = "ASTDY"
 #'       ),
 #'       cmendy = choices_selected(
-#'         choices = variable_choices(ADCM, "CMENDY"),
-#'         selected = "CMENDY"
+#'         choices = variable_choices(ADCM, "AENDY"),
+#'         selected = "AENDY"
 #'       )
 #'     )
 #'   )
 #' )
+#'
 #' \dontrun{
 #' shinyApp(app$ui, app$server)
 #' }